+++ packages/wallets/src/signers/shadow-signers/encodeEd25519PublicKey.ts
